+++ test/MR.Augmenter.Tests/ServiceCollectionExtensionTest.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using Xunit;
 
 namespace MR.Augmenter
@@ -17,16 +18,24 @@ namespace MR.Augmenter
 				{
 					c.Add("Bar", (_, __) => "bar");
 				});
-				config.Build();
+			});
+
+			services.Configure<AugmenterConfiguration>(config =>
+			{
+				config.Configure<TestModel1>(c =>
+				{
+					c.Add("Bar2", (_, __) => "bar2");
+				});
 			});
 
 			var provider = services.BuildServiceProvider();
 			var augmenter = provider.GetRequiredService<IAugmenter>();
-			var configuration = provider.GetRequiredService<AugmenterConfiguration>();
+			var configuration = provider.GetRequiredService<IOptions<AugmenterConfiguration>>().Value;
 
 			configuration.TypeConfigurations.Should()
 				.HaveCount(1).And
 				.Subject.First().Type.Should().Be(typeof(TestModel1));
+			configuration.TypeConfigurations.First().Augments.Should().HaveCount(2);
 		}
 	}
 }